+++ src/middlewares/permissions.js
@@ -1,15 +1,27 @@
 const PermissionsChecker = require('../services/permissions-checker');
 const httpError = require('http-errors');
 const logger = require('../services/logger');
+const ConfigStore = require('../services/config-store');
 
 const getRenderingIdFromUser = user => user.renderingId;
 
-function createCheckPermission(environmentSecret, collectionName) {
-  function checkPermission(permissionName) {
+class PermissionMiddlewareCreator {
+  constructor(collectionName) {
+    this.collectionName = collectionName;
+    this.configStore = ConfigStore.getInstance();
+  }
+
+  _checkPermission(permissionName) {
     return (request, response, next) => {
+      const environmentSecret = this.configStore.lianaOptions.envSecret;
       const renderingId = getRenderingIdFromUser(request.user);
 
-      return new PermissionsChecker(environmentSecret, renderingId, collectionName, permissionName)
+      return new PermissionsChecker(
+        environmentSecret,
+        renderingId,
+        this.collectionName,
+        permissionName,
+      )
         .perform()
         .then(next)
         .catch((error) => {
@@ -19,24 +31,34 @@ function createCheckPermission(environmentSecret, collectionName) {
     };
   }
 
-  function checkPermissionListAndSearch(request, response, next) {
-    const { searchToEdit } = request.query;
-    const renderingId = getRenderingIdFromUser(request.user);
-    const permissionName = searchToEdit ? 'searchToEdit' : 'list';
+  list() {
+    return (request, response, next) => {
+      const { searchToEdit } = request.query;
+      const permissionName = searchToEdit ? 'searchToEdit' : 'list';
 
-    return new PermissionsChecker(environmentSecret, renderingId, collectionName, permissionName)
-      .perform()
-      .then(next)
-      .catch((error) => {
-        logger.error(error.message);
-        next(httpError(403));
-      });
+      return this._checkPermission(permissionName)(request, response, next);
+    };
   }
 
-  return {
-    checkPermission,
-    checkPermissionListAndSearch,
-  };
+  export() {
+    return this._checkPermission('export');
+  }
+
+  details() {
+    return this._checkPermission('show');
+  }
+
+  create() {
+    return this._checkPermission('create');
+  }
+
+  update() {
+    return this._checkPermission('update');
+  }
+
+  delete() {
+    return this._checkPermission('delete');
+  }
 }
 
-module.exports = { createCheckPermission };
+module.exports = PermissionMiddlewareCreator;