+++ src/routes/resources.js
@@ -4,16 +4,20 @@ const ResourceSerializer = require('../serializers/resource');
 const ResourceDeserializer = require('../deserializers/resource');
 const CSVExporter = require('../services/csv-exporter');
 const ParamsFieldsDeserializer = require('../deserializers/params-fields');
-const { createCheckPermission } = require('../middlewares/permissions');
+const PermissionMiddlewareCreator = require('../middlewares/permissions');
+const ConfigStore = require('../services/config-store');
 
-module.exports = function (app, model, Implementation, integrator, opts) {
+const configStore = ConfigStore.getInstance();
+
+module.exports = function Resources(app, model) {
+  const { Implementation, integrator, lianaOptions } = configStore;
   const modelName = Implementation.getModelName(model);
 
-  this.list = function (request, response, next) {
+  this.list = (request, response, next) => {
     const params = request.query;
     const fieldsPerModel = new ParamsFieldsDeserializer(params.fields).perform();
 
-    return new Implementation.ResourcesGetter(model, opts, params)
+    return new Implementation.ResourcesGetter(model, lianaOptions, params)
       .perform()
       .then((results) => {
         const records = results[0];
@@ -24,7 +28,6 @@ module.exports = function (app, model, Implementation, integrator, opts) {
           model,
           records,
           integrator,
-          opts,
           null,
           fieldsSearched,
           params.search,
@@ -37,19 +40,20 @@ module.exports = function (app, model, Implementation, integrator, opts) {
       .catch(next);
   };
 
-  this.count = function (request, response, next) {
+  this.count = (request, response, next) => {
     const params = request.query;
 
-    return new Implementation.ResourcesGetter(model, opts, params)
+    return new Implementation.ResourcesGetter(model, lianaOptions, params)
       .count()
       .then(count => response.send({ count }))
       .catch(next);
   };
 
-  this.exportCSV = function (request, response, next) {
+  this.exportCSV = (request, response, next) => {
     const params = request.query;
-    const recordsExporter = new Implementation.RecordsExporter(
-      model, opts,
+    const recordsExporter = new Implementation.ResourcesExporter(
+      model,
+      lianaOptions,
       params,
     );
     return new CSVExporter(params, response, modelName, recordsExporter)
@@ -57,27 +61,29 @@ module.exports = function (app, model, Implementation, integrator, opts) {
       .catch(next);
   };
 
-  this.get = function (request, response, next) {
-    return new Implementation.ResourceGetter(model, request.params)
-      .perform()
-      .then(record => new ResourceSerializer(
-        Implementation, model, record,
-        integrator, opts,
-      ).perform())
-      .then((record) => {
-        response.send(record);
-      })
-      .catch(next);
-  };
+  this.get = (request, response, next) => new Implementation.ResourceGetter(model, request.params)
+    .perform()
+    .then(record => new ResourceSerializer(
+      Implementation,
+      model,
+      record,
+      integrator,
+    ).perform())
+    .then((record) => {
+      response.send(record);
+    })
+    .catch(next);
 
-  this.create = function (request, response, next) {
+  this.create = (request, response, next) => {
     new ResourceDeserializer(Implementation, model, request.body, true, {
       omitNullAttributes: true,
     }).perform()
       .then(params => new Implementation.ResourceCreator(model, params).perform())
       .then(record => new ResourceSerializer(
-        Implementation, model, record,
-        integrator, opts,
+        Implementation,
+        model,
+        record,
+        integrator,
       ).perform())
       .then((record) => {
         response.send(record);
@@ -85,25 +91,27 @@ module.exports = function (app, model, Implementation, integrator, opts) {
       .catch(next);
   };
 
-  this.update = function (request, response, next) {
+  this.update = (request, response, next) => {
     new ResourceDeserializer(Implementation, model, request.body, false)
       .perform()
       .then((record) => {
         new Implementation.ResourceUpdater(model, request.params, record)
           .perform()
-          .then(record => new ResourceSerializer(
-            Implementation, model, record,
-            integrator, opts,
+          .then(updatedRecord => new ResourceSerializer(
+            Implementation,
+            model,
+            updatedRecord,
+            integrator,
           ).perform())
-          .then((record) => {
-            response.send(record);
-            return record;
+          .then((updatedRecord) => {
+            response.send(updatedRecord);
+            return updatedRecord;
           })
           .catch(next);
       });
   };
 
-  this.remove = function (request, response, next) {
+  this.remove = (request, response, next) => {
     new Implementation.ResourceRemover(model, request.params)
       .perform()
       .then(() => {
@@ -112,52 +120,49 @@ module.exports = function (app, model, Implementation, integrator, opts) {
       .catch(next);
   };
 
-  const {
-    checkPermission,
-    checkPermissionListAndSearch,
-  } = createCheckPermission(opts.envSecret, modelName);
+  const permissionMiddlewareCreator = new PermissionMiddlewareCreator(modelName);
 
-  this.perform = function () {
+  this.perform = () => {
     app.get(
-      `${path.generate(modelName, opts)}.csv`,
+      `${path.generate(modelName, lianaOptions)}.csv`,
       auth.ensureAuthenticated,
-      checkPermission('export'),
+      permissionMiddlewareCreator.export(),
       this.exportCSV,
     );
     app.get(
-      path.generate(modelName, opts),
+      path.generate(modelName, lianaOptions),
       auth.ensureAuthenticated,
-      checkPermissionListAndSearch,
+      permissionMiddlewareCreator.list(),
       this.list,
     );
     app.get(
-      path.generate(`${modelName}/count`, opts),
+      path.generate(`${modelName}/count`, lianaOptions),
       auth.ensureAuthenticated,
-      checkPermissionListAndSearch,
+      permissionMiddlewareCreator.list(),
       this.count,
     );
     app.get(
-      path.generate(`${modelName}/:recordId`, opts),
+      path.generate(`${modelName}/:recordId`, lianaOptions),
       auth.ensureAuthenticated,
-      checkPermission('show'),
+      permissionMiddlewareCreator.details(),
       this.get,
     );
     app.post(
-      path.generate(modelName, opts),
+      path.generate(modelName, lianaOptions),
       auth.ensureAuthenticated,
-      checkPermission('create'),
+      permissionMiddlewareCreator.create(),
       this.create,
     );
     app.put(
-      path.generate(`${modelName}/:recordId`, opts),
+      path.generate(`${modelName}/:recordId`, lianaOptions),
       auth.ensureAuthenticated,
-      checkPermission('update'),
+      permissionMiddlewareCreator.update(),
       this.update,
     );
     app.delete(
-      path.generate(`${modelName}/:recordId`, opts),
+      path.generate(`${modelName}/:recordId`, lianaOptions),
       auth.ensureAuthenticated,
-      checkPermission('delete'),
+      permissionMiddlewareCreator.delete(),
       this.remove,
     );
   };