+++ src/services/exposed/record-getter.js
@@ -1,0 +1,10 @@
+const AbstractRecordService = require('./abstract-records-service');
+
+class RecordGetter extends AbstractRecordService {
+  get(recordId) {
+    return new this.Implementation.ResourceGetter(this.model, { recordId })
+      .perform();
+  }
+}
+
+module.exports = RecordGetter;