+++ src/services/exposed/record-remover.js
@@ -1,0 +1,10 @@
+const AbstractRecordService = require('./abstract-records-service');
+
+class RecordRemover extends AbstractRecordService {
+  remove(recordId) {
+    return new this.Implementation.ResourceRemover(this.model, { recordId })
+      .perform();
+  }
+}
+
+module.exports = RecordRemover;